+++ App.js
@@ -6,6 +6,7 @@ import RegisterPage from "./src/screens/RegisterPage";
 import ProfilePage from "./src/screens/ProfilePage";
 import EditPage from "./src/screens/EditPage";
 
+
 const Stack = createNativeStackNavigator();
 
 function App() {
@@ -17,7 +18,7 @@ function App() {
           name="Login"
           component={LoginPage}
         />
-        <Stack.Screen options={{ headerShown: false }} name="Main" component={MainPage} />
+        <Stack.Screen name="Main" component={MainPage} />
         <Stack.Screen name="Register" component={RegisterPage} />
         <Stack.Screen name="Profile" component={ProfilePage} />
         <Stack.Screen name="Edit" component={EditPage} />
