+++ package.json
@@ -9,13 +9,16 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@expo/config-plugins": "^5.0.4",
     "@expo/webpack-config": "^0.17.2",
+    "@react-native-async-storage/async-storage": "^1.17.11",
     "@react-navigation/bottom-tabs": "^6.5.4",
     "@react-navigation/native": "^6.1.3",
     "@react-navigation/native-stack": "^6.9.9",
     "@types/react": "~18.0.24",
     "@types/react-native": "~0.70.6",
-    "expo": "~47.0.12",
+    "expo": "^47.0.13",
+    "expo-image-picker": "~14.0.2",
     "expo-status-bar": "~1.4.2",
     "firebase": "^9.17.1",
     "galio-framework": "^0.8.0",
@@ -27,11 +30,11 @@
     "react-native-safe-area-context": "4.4.1",
     "react-native-screens": "~3.18.0",
     "react-native-web": "~0.18.9",
-    "typescript": "^4.6.3",
-    "expo-image-picker": "~14.0.2"
+    "typescript": "^4.6.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9"
+    "@babel/core": "^7.12.9",
+    "@expo/config": "^7.0.3"
   },
   "private": true
 }
